Unique ID Management Project (Aug 11 - Feb 12) - IBM's TGMC Competition
Web Application which maintains a record of personal details of every
citizen in India. It also provides other benefits such as booking
tickets, filing complaints, paying taxes etc. online. The purpose of UID
is to help citizens avail of the various government services in a fast
and simple manner as well as provide a complete record of all citizens
as required by the government authorities.